+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - Fix the farming wand, add additional safety for unusable materials.
  - Allow tesseract to pass through glass
  - Fix interaction with Essentials signs and Wands interfering with each other.
+ - Add dropped wands to the dynmap (!)
 
 ## 2.8.0